+++ odoo/local-src/delivery_rounds/models/__init__.py
@@ -3,3 +3,5 @@ from . import zone
 from . import instance
 from . import partner
 from . import picking
+from . import delivery_carrier
+from . import sale_order